Refactor feiwen module boundaries
