+++ app/ai-chat/src/components.rs
@@ -6,7 +6,6 @@ pub mod chat_form;
 pub mod delete_confirm;
 pub mod hotkey_input;
 pub mod message;
-pub(crate) mod search_list;
 pub(crate) mod title_bar_menu;
 
 pub(crate) fn init(cx: &mut App) {
