+++ app/ai-chat/src/features/home/search_list.rs
