+++ app/feiwen/src/app.rs
@@ -1,0 +1,4 @@
+pub(crate) mod titlebar;
+mod workspace;
+
+pub(crate) use workspace::{RouterType, Workspace, WorkspaceEvent, WorkspaceView};
