+++ app/feiwen/src/features.rs
@@ -1,6 +1,2 @@
-mod fetch;
-mod query;
-pub(crate) mod titlebar;
-mod workspace;
-
-pub(crate) use workspace::{Workspace, WorkspaceView};
+pub(crate) mod fetch;
+pub(crate) mod query;
