+++ app/feiwen/src/features/query/advanced.rs
@@ -1,6 +1,8 @@
+mod components;
 mod options;
 mod render;
 mod sort;
+mod spec;
 mod state;
 
 pub(crate) use options::QueryOptions;
